+++ sdk/communication/communication-short-codes/samples/v1/javascript/createAndSubmitProgramBrief.js
@@ -82,6 +82,7 @@ async function main() {
         totalMonthlyVolume: 10000,
         monthlyAverageMessagesFromUser: 1,
         monthlyAverageMessagesToUser: 3,
+        estimatedRampUpTimeInDays: 50,
         isSpiky: true,
         spikeDetails:
           "Higher traffic expected around major shopping holidays, most notably Black Friday and Memorial Day."