+++ sdk/communication/communication-short-codes/src/models.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 import { OperationOptions } from "@azure/core-http";
+import { ShortCodesGetShortCodesOptionalParams } from ".";
 
 /**
  * Additional options for the delete US Program Brief request.
@@ -17,6 +18,16 @@ export interface GetUSProgramBriefOptions extends OperationOptions {}
  * Additional options for the Submit US Program Brief request.
  */
 export interface SubmitUSProgramBriefOptions extends OperationOptions {}
+
+/**
+ * Additional options for the List Short Codes request.
+ */
+export interface ListShortCodesOptions extends ShortCodesGetShortCodesOptionalParams {}
+
+/**
+ * Additional options for the List US Program Brief request.
+ */
+export interface ListUSProgramBriefsOptions extends OperationOptions {}
 
 export {
   ShortCode,