+++ sdk/communication/communication-short-codes/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
-## 1.0.0-beta.1 (2021-11-09)
+## 1.0.0-beta.1 (Unreleased)
 
-- Test Release Pipeline
+The first preview of the Azure Communication Short Codes Client has the following features:
+
+- Management of US Program Briefs
+  - Create, Update, Get, List & Delete
+- US Program Brief submission for review to opt for a Short Code
+- List owned Short Codes